Addresses styling and layout concerns after review
