+++ personal-portfolio/src/components/Home.css
@@ -1,0 +1,18 @@
+.headline {
+    text-align: center;
+    border: #888;
+  }
+  
+.profile {
+    margin-left: auto;
+    margin-right: auto;
+    margin-top: 5em;
+    border-radius: 100%;
+    width: 350px;
+    display: block;
+  }
+
+.socials {
+    text-align: center;
+    margin: 10px;
+  }
